+++ eras/mary/impl/src/Cardano/Ledger/Mary/Scripts.hs
@@ -7,11 +7,14 @@ module Cardano.Ledger.Mary.Scripts (
 )
 where
 
-import Cardano.Ledger.Allegra.Scripts (Timelock, translateTimelock)
+import Cardano.Ledger.Allegra.Scripts
 import Cardano.Ledger.Core
-import Cardano.Ledger.Crypto (Crypto)
+import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
 import Cardano.Ledger.Mary.Era (MaryEra)
-import Cardano.Ledger.Shelley.Scripts (nativeMultiSigTag)
+import Cardano.Ledger.Shelley.Scripts (
+  ShelleyEraScript (..),
+  nativeMultiSigTag,
+ )
 
 -- | Since Timelock scripts are a strictly backwards compatible extension of
 -- MultiSig scripts, we can use the same 'scriptPrefixTag' tag here as we did
@@ -27,3 +30,27 @@ instance Crypto c => EraScript (MaryEra c) where
   getNativeScript = Just
 
   fromNativeScript = id
+
+instance Crypto c => ShelleyEraScript (MaryEra c) where
+  {-# SPECIALIZE instance ShelleyEraScript (MaryEra StandardCrypto) #-}
+
+  mkRequireSignature = mkRequireSignatureTimelock
+  getRequireSignature = getRequireSignatureTimelock
+
+  mkRequireAllOf = mkRequireAllOfTimelock
+  getRequireAllOf = getRequireAllOfTimelock
+
+  mkRequireAnyOf = mkRequireAnyOfTimelock
+  getRequireAnyOf = getRequireAnyOfTimelock
+
+  mkRequireMOf = mkRequireMOfTimelock
+  getRequireMOf = getRequireMOfTimelock
+
+instance Crypto c => AllegraEraScript (MaryEra c) where
+  {-# SPECIALIZE instance AllegraEraScript (MaryEra StandardCrypto) #-}
+
+  mkTimeStart = mkTimeStartTimelock
+  getTimeStart = getTimeStartTimelock
+
+  mkTimeExpire = mkTimeExpireTimelock
+  getTimeExpire = getTimeExpireTimelock