+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/Imp.hs
@@ -7,6 +7,7 @@
 
 module Test.Cardano.Ledger.Mary.Imp (spec) where
 
+import Cardano.Ledger.Allegra.Scripts
 import Cardano.Ledger.Mary.Core
 import Cardano.Ledger.Shelley.Rules (ShelleyUtxoPredFailure, ShelleyUtxowPredFailure)
 import qualified Test.Cardano.Ledger.Allegra.Imp as AllegraImp
@@ -18,6 +19,7 @@ import Test.Cardano.Ledger.Shelley.ImpTest (withImpState)
 spec ::
   forall era.
   ( MaryEraImp era
+  , AllegraEraScript era
   , InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   ) =>