+++ eras/babbage/impl/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.8.0.1
+## 1.8.1.0
 
-*
+* Add `AllegraEraScript` and `ShelleyEraScript` instances for `BabbageEra`
 
 ## 1.8.0.0